* Restructure the AlsaDriver's note-off handling to ensure that outgoing   events are always queued in order.  This is not generally necessary for   ALSA sequencer queues, but it's vital for soft synths, and it allows us   to simplify things by always using our internal note-off queue instead   of sometimes using the ALSA sequencer queue's note-off generation and   sometimes our own.  (And sometimes both, as we were doing before.) * More soft-synth work.  We can actually play through them now, but there's   still no way to select them in the first place...
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@5725 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/DSSIPluginInstance.h
@@ -66,7 +66,7 @@ public:
     virtual bool isBypassed() const { return m_bypassed; }
     virtual void setBypassed(bool bypassed) { m_bypassed = bypassed; }
 
-    virtual void setIdealChannelCount(int channels); // may re-instantiate
+    virtual void setIdealChannelCount(size_t channels); // may re-instantiate
 
 protected:
     // To be constructed only by DSSIPluginFactory
@@ -120,8 +120,8 @@ protected:
     sample_t                **m_inputBuffers;
     sample_t                **m_outputBuffers;
     bool                      m_ownBuffers;
-    int                       m_idealChannelCount;
-    int                       m_outputBufferCount;
+    size_t                    m_idealChannelCount;
+    size_t                    m_outputBufferCount;
     size_t                    m_sampleRate;
     
     bool                      m_bypassed;
